requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
decorator==4.4.2
dnspython==2.4.2
fastapi==0.103.2
ffmpeg==1.4
filelock==3.12.4
future==0.18.3
h11==0.14.0
idna==3.4
imageio==2.31.5
imageio-ffmpeg==0.4.9
Jinja2==3.1.2
llvmlite==0.41.0
MarkupSafe==2.1.3
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.1
numba==0.58.0
numpy==1.25.2
oauthlib==3.2.2
openai-whisper @ git+https://github.com/openai/whisper.git@0a60fcaa9b86748389a656aa013c416030287d47
pika==1.3.2
Pillow==10.0.1
proglog==0.1.10
pydantic==2.4.2
pydantic_core==2.10.1
pymongo==4.5.0
python-multipart==0.0.6
regex==2023.8.8
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tiktoken==0.3.3
torch==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.5
uvicorn==0.23.2
